+++ src/core/client/api/consts/utilityConst.ts
@@ -3,9 +3,9 @@ import { loadAnimation } from '@AthenaClient/systems/animations';
 import { getClosestPlayer, getClosestVehicle } from '@AthenaClient/utility/closest';
 import { drawMissionText } from '@AthenaClient/utility/missionText';
 import { loadModel } from '@AthenaClient/utility/model';
-import { UID } from '@AthenaClient/utility/uid';
 import { switchToMultiSecondpart } from '@AthenaClient/utility/switch';
 import { isEntityBlockingPosition } from '@AthenaClient/utility/isEntityBlocking';
+import { UID } from '@AthenaShared/utility/uid';
 
 export const utilityConst = {
     loadModel,