+++ src/core/server/api/consts/constUtility.ts
@@ -3,6 +3,7 @@ import { hashConst } from './constHash';
 import { vectorConst } from './constVector';
 import { isFlagEnabled } from '../../../shared/utility/flags';
 import { getClosestPlayer, getClosestVehicle } from '../../utility/closest';
+import { UID } from '@AthenaShared/utility/uid';
 
 export const utilityConst = {
     hash: hashConst,
@@ -11,4 +12,5 @@ export const utilityConst = {
     isFlagEnabled,
     getClosestPlayer,
     getClosestVehicle,
+    uid: UID,
 };