+++ src/core/server/systems/defaults/ammo.ts
@@ -3,7 +3,10 @@ import * as alt from 'alt-server';
 import { PluginSystem } from '../plugins';
 import { Athena } from '@AthenaServer/api/athena';
 import { getWeaponMap } from '@AthenaShared/information/weaponList';
-import { StoredItem } from '@AthenaShared/interfaces/item';
+import { StoredItem, WeaponInfo } from '@AthenaShared/interfaces/item';
+import { SYSTEM_EVENTS } from '@AthenaShared/enums/system';
+import { sha256Random } from '@AthenaServer/utility/encryption';
+import { deepCloneArray } from '@AthenaShared/utility/deepCopy';
 
 /**
  * THIS IS A DEFAULT SYSTEM.
@@ -14,26 +17,71 @@ import { StoredItem } from '@AthenaShared/interfaces/item';
  * COPY THE CODE AND REMAKE IT AS A PLUGIN IF YOU WANT TO MAKE CHANGES.
  */
 
-type AmmoOrWeapon = { ammo: number; hash?: number };
-
 const SYSTEM_NAME = 'Ammunition Items';
+const hashes = {};
 
 let enabled = true;
 
 const Internal = {
-    combineData(item1: StoredItem<AmmoOrWeapon>, item2: StoredItem<AmmoOrWeapon>): Required<AmmoOrWeapon> {
-        let data: { hash?: number; ammo: number } = { ammo: 0 };
+    handleUpdate(player: alt.Player, uid: string, hash: number, ammo: number) {
+        if (!hashes[player.id] || hashes[player.id] !== uid) {
+            return;
+        }
+
+        const data = Athena.document.character.get(player);
+        if (typeof data === 'undefined') {
+            return;
+        }
+
+        const toolbar = deepCloneArray<StoredItem<WeaponInfo>>(data.toolbar);
+        const toolbarIndex = toolbar.findIndex((x) => x.isEquipped && x.data.hash && x.data.hash === hash);
+        if (toolbarIndex <= -1) {
+            return;
+        }
+
+        if (toolbar[toolbarIndex].data.ammo === ammo) {
+            return;
+        }
+
+        toolbar[toolbarIndex].data.ammo = ammo;
+        Athena.document.character.set(player, 'toolbar', toolbar, true);
+    },
+    weaponUpdateTick() {
+        const players = Athena.get.players.onlineWithWeapons();
+        for (let armedPlayer of players) {
+            if (!armedPlayer || !armedPlayer.valid) {
+                continue;
+            }
+
+            const uid = Athena.utility.uid.generate();
+            hashes[armedPlayer.id] = uid;
+            armedPlayer.emit(SYSTEM_EVENTS.PLAYER_EMIT_AMMUNITION_UPDATE, uid, armedPlayer.currentWeapon);
+        }
+    },
+    combineData(item1: StoredItem<WeaponInfo>, item2: StoredItem<WeaponInfo>): Required<WeaponInfo> {
+        let data: WeaponInfo = {
+            hash: undefined,
+            ammo: 0,
+        };
 
         if (typeof item1.data.hash !== 'undefined') {
             data.hash = item1.data.hash;
+        }
+
+        if (typeof item1.data.components !== 'undefined') {
+            data.components = item1.data.components;
         }
 
         if (typeof item2.data.hash !== 'undefined') {
             data.hash = item2.data.hash;
         }
 
+        if (typeof item2.data.components !== 'undefined') {
+            data.components = item2.data.components;
+        }
+
         data.ammo = item1.data.ammo + item2.data.ammo;
-        return data as Required<AmmoOrWeapon>;
+        return data as Required<WeaponInfo>;
     },
     async init() {
         if (!enabled) {
@@ -63,14 +111,17 @@ const Internal = {
                     quantity: 1,
                     data: Internal.combineData,
                 },
+                sound: 'weapon_load',
             });
         });
 
+        alt.setInterval(Internal.weaponUpdateTick, 500);
+        alt.onClient(SYSTEM_EVENTS.PLAYER_EMIT_AMMUNITION_UPDATE, Internal.handleUpdate);
         alt.log(`~lc~Default System: ~g~${SYSTEM_NAME}`);
     },
 };
 
-export const DefaultAmmoItemsSystem = {
+export const DefaultAmmoSystem = {
     disable: () => {
         enabled = false;
         alt.log(`~y~Default ${SYSTEM_NAME} Turned Off`);